Gibberish strings were not being looked up with new cache changes.
Not sure how this made it through, but it wasn't doing lookups at all
due to a type problem in the built out cache as well as the lookup
key.

We've written unit tests in our application for this, but they're not
yet available for the plugin.  Hopefully we can figure out how to test
plugins in isolation a bit better in the near future.